project/requirements.txt
# Core runtime
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
flask>=2.3,<4
nltk>=3.8,<4
python-dotenv>=1.0,<2
openai>=1.40,<2
pydantic>=2.8,<3

# Testing
pytest>=8.0,<9
httpx>=0.27,<1

# Optional legacy-local inference stack
# torch>=2.2
# transformers>=4.40
# sentencepiece>=0.1.99
# joblib>=1.4
